+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.12.4](https://github.com/eclipse/paho.mqtt.rust/compare/v0.12.3..v0.12.4) - 2024-05-19
+
+- Fixes for topic matching:
+    - `TopicMatcher`
+        - Fixed a number of corner cases
+        - Iterator optimized
+        - Added `prune()` and `shrink_to_fit()`, and `get_key_value()`
+    - `TopicFilter` fixed corner cases
+    - Added stand-alone `topic_matches()` and `topic_matches_iter()` functions from [PR #228](https://github.com/eclipse/paho.mqtt.rust/pull/228)
+
+
 ## [v0.12.3](https://github.com/eclipse/paho.mqtt.rust/compare/v0.12.2..v0.12.3) - 2023-10-25
 
 - The -sys crate now wraps Paho C v1.3.13, fixing several issues, including crashes on reconnect callbacks.